feat(last): add higher-order lettable version of last
