+++ src/operator/last.ts
@@ -1,7 +1,5 @@
 import { Observable } from '../Observable';
-import { Operator } from '../Operator';
-import { Subscriber } from '../Subscriber';
-import { EmptyError } from '../util/EmptyError';
+import { last as higherOrder } from '../operators';
 
 /* tslint:disable:max-line-length */
 export function last<T, S extends T>(this: Observable<T>,
@@ -45,94 +43,5 @@ export function last<T>(this: Observable<T>,
 export function last<T, R>(this: Observable<T>, predicate?: (value: T, index: number, source: Observable<T>) => boolean,
                            resultSelector?: ((value: T, index: number) => R) | void,
                            defaultValue?: R): Observable<T | R> {
-  return this.lift(new LastOperator(predicate, resultSelector, defaultValue, this));
-}
-
-class LastOperator<T, R> implements Operator<T, R> {
-  constructor(private predicate?: (value: T, index: number, source: Observable<T>) => boolean,
-              private resultSelector?: ((value: T, index: number) => R) | void,
-              private defaultValue?: any,
-              private source?: Observable<T>) {
-  }
-
-  call(observer: Subscriber<R>, source: any): any {
-    return source.subscribe(new LastSubscriber(observer, this.predicate, this.resultSelector, this.defaultValue, this.source));
-  }
-}
-
-/**
- * We need this JSDoc comment for affecting ESDoc.
- * @ignore
- * @extends {Ignored}
- */
-class LastSubscriber<T, R> extends Subscriber<T> {
-  private lastValue: T | R;
-  private hasValue: boolean = false;
-  private index: number = 0;
-
-  constructor(destination: Subscriber<R>,
-              private predicate?: (value: T, index: number, source: Observable<T>) => boolean,
-              private resultSelector?: ((value: T, index: number) => R) | void,
-              private defaultValue?: any,
-              private source?: Observable<T>) {
-    super(destination);
-    if (typeof defaultValue !== 'undefined') {
-      this.lastValue = defaultValue;
-      this.hasValue = true;
-    }
-  }
-
-  protected _next(value: T): void {
-    const index = this.index++;
-    if (this.predicate) {
-      this._tryPredicate(value, index);
-    } else {
-      if (this.resultSelector) {
-        this._tryResultSelector(value, index);
-        return;
-      }
-      this.lastValue = value;
-      this.hasValue = true;
-    }
-  }
-
-  private _tryPredicate(value: T, index: number) {
-    let result: any;
-    try {
-      result = this.predicate(value, index, this.source);
-    } catch (err) {
-      this.destination.error(err);
-      return;
-    }
-    if (result) {
-      if (this.resultSelector) {
-        this._tryResultSelector(value, index);
-        return;
-      }
-      this.lastValue = value;
-      this.hasValue = true;
-    }
-  }
-
-  private _tryResultSelector(value: T, index: number) {
-    let result: any;
-    try {
-      result = (<any>this).resultSelector(value, index);
-    } catch (err) {
-      this.destination.error(err);
-      return;
-    }
-    this.lastValue = result;
-    this.hasValue = true;
-  }
-
-  protected _complete(): void {
-    const destination = this.destination;
-    if (this.hasValue) {
-      destination.next(this.lastValue);
-      destination.complete();
-    } else {
-      destination.error(new EmptyError);
-    }
-  }
+  return higherOrder(predicate, resultSelector as any, defaultValue)(this);
 }
